chore: Refactored functions related to variant images as a hook
